Global code revision. Termux branch termination
