+++ WOLSAREnglishClubTelegramBot.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "wolsarenglishclubbot", "wolsarenglishclubbot.csproj", "{C279C357-EB9D-EE33-83A8-0BD5D5F3F55D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C279C357-EB9D-EE33-83A8-0BD5D5F3F55D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C279C357-EB9D-EE33-83A8-0BD5D5F3F55D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C279C357-EB9D-EE33-83A8-0BD5D5F3F55D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C279C357-EB9D-EE33-83A8-0BD5D5F3F55D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
