+++ wolsarenglishclubbot.csproj
@@ -8,8 +8,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.3" />
-    <PackageReference Include="Telegram.Bot" Version="22.4.4" />
+    <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.4" />
+    <PackageReference Include="Telegram.Bot" Version="22.5.2-dev.3" />
   </ItemGroup>
 
 </Project>
